Moved components to manager folder
